publishers: badges: Update mechanism for available/unavailable
As we now have badges which may not be relevant and not attained then we
need to make sure not to list these are not attained when it isn't
actually possible. An example of this is recipient org-ids present in
data from a publisher which only publishes grants where the recipients
are individuals.

Fixes: https://github.com/ThreeSixtyGiving/registry/issues/136